Add methods to log in and checking the logged in status
First thought of using Axios but soon noticed that maybe the original
implementation of the Strava2kilometrikisa was required to get all
the data to by pass the CORS verification. Therefore ended up using
the old Curl package and simply wrapping it to promise api.

We will see if there is any use for Axios or if it could just be removed
from dependencies...