Catch most common Axios errors and throw domain specific error instead
For example interpret 404 errors as not found errors and report
login failures with specific errors so that the user can more easily
figure what is going on in each error case.

Did not bother catching and swallowing all possible Axios errors.
Probably better not to obfuscate those rare error cases and instead let
client user to deal with them if they want to.