Add method to get user's log entries for competition
Returns all user's entries for the given competition and year. There
seems to be some kind of a bug in kilometrikisa service though which
causes its API to return only 0 values. This might be because at the time
of writing new competition has been created but is not active yet.